requirements.txt
uvicorn[standard]
gunicorn
fastapi~=0.61.2
starlette~=0.13.6
cupy~=8.1.0
aiofiles~=0.6.0
Pillow~=8.0.1
redis~=3.5.3
